Make line-index an external lib
